Add cleanup for orphaned hotlinks, add/remove hotlinks on edit #136

+++ internal/configuration/database/hotlinks.go
@@ -27,9 +27,23 @@ func GetHotlink(id string) (string, bool) {
 	return rowResult.FileId, true
 }
 
+// GetAllHotlinks returns an array with all hotlink ids
+func GetAllHotlinks() []string {
+	var ids []string
+	rows, err := sqliteDb.Query("SELECT Id FROM Hotlinks")
+	helper.Check(err)
+	defer rows.Close()
+	for rows.Next() {
+		rowData := schemaHotlinks{}
+		err = rows.Scan(&rowData.Id)
+		helper.Check(err)
+		ids = append(ids, rowData.Id)
+	}
+	return ids
+}
+
 // SaveHotlink stores the hotlink associated with the file in the database
 func SaveHotlink(file models.File) {
-
 	newData := schemaHotlinks{
 		Id:     file.HotlinkId,
 		FileId: file.Id,
@@ -40,8 +54,11 @@ func SaveHotlink(file models.File) {
 	helper.Check(err)
 }
 
-// DeleteHotlink deletes a hotlink with the given ID
+// DeleteHotlink deletes a hotlink with the given hotlink ID
 func DeleteHotlink(id string) {
+	if id == "" {
+		return
+	}
 	_, err := sqliteDb.Exec("DELETE FROM Hotlinks WHERE Id = ?", id)
 	helper.Check(err)
 }
